requirements.txt
django
djangorestframework
pillow
simplejwt